requirements.txt
fastapi
uvicorn[standard]
pandas
python-dateutil
boto3
python-dotenv
